refactor: move backgroundscheduler into own module + some small stuff
background scheduler is now in its own module; also scheduler jobs now coalesce,
have IDs, and the publications job is scheduled to 3 a.m. EST
