+++ runserver.py
@@ -6,9 +6,6 @@
 
 import os
 
-import atexit
-from datetime import date, datetime
-from apscheduler.schedulers.background import BackgroundScheduler
 from config.settings import DOWNLOADS_DIR_PATH
 
 from tornado.httpserver import HTTPServer
@@ -23,6 +20,7 @@ from pathlib2 import Path
 
 from app_and_db import app, db
 from utility import random_string, oe_password_validator
+from helpers.background_scheduler import start_background_scheduler
 from pages.publications import load_default_publications_and_papers, download_and_update_papers_and_bibtex
 from postgres.users import Role, User, add_user, create_role
 from pages.neem_overview import download_neem_files, load_default_overview_files
@@ -36,11 +34,6 @@ USER_ROLES = [
     'user',
     'editor'
 ]
-
-OVERVIEW_SCHEDULER_JOB_ID = 'overview'
-PUBLICATIONS_SCHEDULER_JOB_ID = 'publications'
-
-BACKGROUND_SCHEDULER = BackgroundScheduler()
 
 def init_app(extra_config_settings={}):
     # Initialize app config settings
@@ -118,7 +111,7 @@ def init_app(extra_config_settings={}):
         download_neem_files()
         download_and_update_papers_and_bibtex()
         # start background jobs for periodic fetching
-        _start_background_scheduler()
+        start_background_scheduler()
     
     app.logger.info("Webapp started.")
     return app
@@ -128,20 +121,6 @@ def _config_is_debug():
     # if environment variable 'EASE_DEBUG' is set to true, then
     # 'DEBUG' in app.config will be set to true by init_app.py
     return 'DEBUG' in app.config and app.config['DEBUG']
-
-
-def _start_background_scheduler():
-    BACKGROUND_SCHEDULER.add_job(func=download_neem_files, trigger="interval", hours=3, coalesce=True, id=OVERVIEW_SCHEDULER_JOB_ID)
-    BACKGROUND_SCHEDULER.add_job(func=download_and_update_papers_and_bibtex, trigger="interval", days=1, next_run_time=_get_tomorrow_3_am_datetime(), coalesce=True, id=PUBLICATIONS_SCHEDULER_JOB_ID)
-    BACKGROUND_SCHEDULER.start()
-
-    # Shut down the scheduler when exiting the app
-    atexit.register(lambda: BACKGROUND_SCHEDULER.shutdown())
-
-
-def _get_tomorrow_3_am_datetime():
-    today = date.today()
-    return datetime(today.year, today.month, today.day, 3, 0, 0)
 
 
 def _run_server():
